Rework how encodings are handled
And replace UTF-8 with UCS-2 as encoding number 0. I was given a font file that uses encoding 0, is clearly using UCS-2, and works correctly in games. So I think all the existing documentation that claims that 0 is UTF-8 is wrong. If you've got a better explanation (ideally with a counterexample), please open an issue!